Port console error step to general
